fix(api-sdk): fix data object in createGroups request
Now the right data object is sent in the createGroups request in the API SDK. If the credentials
parameter is undefined, it will be set to null in the server. The group id is no longer part of the
parameters to create a group, the group ids will be managed on the server.

BREAKING CHANGE: The group id is no longer part of the parameters to create a group when using the
API SDK.

fix #466